Better error messages for syntax errors
Fixes #514

The error messages from the vendored copy of rewrite-clj are not as
helpful as they could be. I often have a hard time finding the location
of a mismatched bracket, since the error message refers to the EOF at
the end of the file, and not the opening paren.

To fix this we allow rewrite-clj to throw exceptions containing a
sequence of errors.

Changes:

- Make rewrite-clj throw ex-info instead of Exception. The ex-data
contains a :findings key, which is a sequence of findings in the format
expected by clj-kondo

- Change rewrite-clj.parser.core/*delimiters* to keep track of a tuple
of [open, close, row, col] rather than just the expected closing
delimiter. The extra three fields allow is to construct more informative
error messages when a parse error occurs.

- If an EOF is encountered when parsing a delimited list {}, [] or (),
throw an exception with two error messages - one for the opening
delimiter, saying that it is unmatched, and one for the end of file
saying that a closing delimiter was expected.

- If a mismatched delimiter is found when parsing a list ({}, [], ())
then throw two error messages - one for the opening bracket and one for
the closing bracket. In both cases, refer to both the line on which the
token is found, and the line on which the corresponding token is found.

Before:

> Unexpected EOF

After:

> Found an opening [ with no matching ]
> Expected a ] to match [ from line 1

Before:

> Unmatched delimiter: }

After:

> Mismatched bracket: found an opening [ and a closing } on line 1
> Mismatched bracket: found an opening [ on line 1 and a closing }