Use an indexing reader when parsing EDN
By passing an indexing reader to the edn parser, line and column
information is produced in the error messages.

    clj-kondo.impl.toolsreader.v1v2v2.clojure.tools.reader.impl.errors/throw-ex

Remove the string formatting here. All of the data required by
->findings is contained in the ex-data. If we don't remove this string
formatting, then the error message shown to the user will be prepended
with `[line x, col y] `.

    clj-kondo.impl.analyzer/->findings

Now that tools.reader is throwing data, we need to update `->findings`
to be able to read that data correctly.

Add unit test to ensure that line data comes out of the parser, and an
integration to ensure that `->findings` works.