Allow ignore unused :as binding.
It is sometimes useful to include the `:as` when destructuring maps because it
can help to identify similarily named functions, but with different inputs,
and thus acts as a form of documentation.

For example:

`(defn foo [{:keys [bar] :as config}] bar)`

Whilst `config` is not used, it is at least documented and easy to spot when
scanning through the code.

This commit, which closes #1016, introduces a new exclude-binding option
called `exclude-destructured-as` which, when set to true, will tell the linter to
ignore any unused :as bindings.

-=david=-