Include namespace and name in unused-refer findings (#1052)
* feat: include ns and name in :unused-referred-var findings

This extends the findings map for :unused-referred-vars to include the
:ns and :name of the var. Intended to support usage in `carve`.

* test: ensure unused-referred-vars include :ns and :name

Adds a test to ensure the unused-referred var :findings include a
namespace and name key.

This was impled in core_test and against run! rather than against lint!,
because lint! parses maps back from the printed output, which won't
include this metadata.