Limit var analysis (#1674)
* Can disable analysis of var-usages

* Can analyze var-definitions shallowly

If you set {:output {:analysis {:var-definitions {:shallow true}}}}, the
names (and optionally metadata and arglists) of vars will be analyzed, but
their bodies will be skipped.

This is useful to analyze the skeleton of a namespace, without getting
any details. It is an expert option, and if you use it, you should not
expect linters to behave correctly anymore.

More specifically, the bodies of these forms are skipped:

- def
- defn
- defrecord
- defmethod
- plumatic schema; fn, def, defn, defmethod, defrecord
- protocol-impls
- fn
- letfn
- fn*
- bound-fn

* Move :analysis config out of :output

* Support running analysis without lint

* Minimize analysis in tools

* Changelog for limiting var analysis

* Revert redundant checks and inline local var

Per PR feedback.