Add new `--report-level` CLI option
This is useful for CI where you may not want to display info-level
notices if you only require folks to fix warnings or above.

The default level is `info`.

With `--report-level error` will also omit the `, warnings: N` part of
the output summary.

If the report level is not recognised, prints the help.

Fixes #2410